refactor: Renamed CheckBox.tsx to Checkbox.tsx
