+++ components/ui/components/design/atoms/inputs/index.ts
@@ -1,5 +1,5 @@
 export { ComboBox } from "./ComboBox";
-export { Checkbox } from "./CheckBox";
+export { Checkbox } from "./Checkbox";
 export { Input } from "./Input";
 export { Button } from "./Button";
 export { MenuButton } from "./MenuButton";
